:white_check_mark: Add support for TEST_BUILD environment variable
In the `test_build` function, added support for the TEST_BUILD environment variable. Now, if the variable is set to True, the `test_build` function will run; otherwise, it will skip the test and output a related message. Additionally, introduced a new pytest fixture `if_test_build` to check if the test is enabled.